% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R, R/allGenerics.R
\docType{class}
\name{Agent-class}
\alias{Agent-class}
\alias{Agent}
\alias{getLocation,Agent-method}
\alias{getVelocity,Agent-method}
\alias{getActive,Agent-method}
\alias{setLocation,Agent-method}
\alias{setVelocity,Agent-method}
\alias{setActive,Agent-method}
\title{An S4 class to represent an agent.}
\usage{
\S4method{getLocation}{Agent}(object)

\S4method{getVelocity}{Agent}(object)

\S4method{getActive}{Agent}(object)

\S4method{setLocation}{Agent}(object, location)

\S4method{setVelocity}{Agent}(object, velocity)

\S4method{setActive}{Agent}(object, active)
}
\arguments{
\item{object}{An object of class "Agent"}

\item{location}{A numerical vector of length 2 specifying the x- and
y-position of the object [Default: c(0, 0)].}

\item{velocity}{A numerical vector of length 2 specifying the velocity in the
x- and y-direction of the object [Default: c(0, 0)].}

\item{active}{A length-one logical vector indicating whether the object is
active (TRUE) or not (FALSE) [Default: TRUE].}
}
\description{
An S4 class to represent an agent.
}
\section{Methods (by generic)}{
\itemize{
\item \code{getLocation}: Retreive the location slot of an object of class "Agent".

\item \code{getVelocity}: Retreive the velocity slot of an object of class "Agent".

\item \code{getActive}: Retreive the active slot of an object of class "Agent".

\item \code{setLocation}: Assign the location slot of an object of class "Agent".

\item \code{setVelocity}: Assign the velocity slot of an object of class "Agent".

\item \code{setActive}: Assign the active slot of an object of class "Agent".
}}

\section{Slots}{

\describe{
\item{\code{location}}{A numerical vector of length 2 specifying the x- and
 y-position of the object of class \linkS4class{Agent}
[Default: c(0, 0)].}

\item{\code{velocity}}{A numeric vector for specifying the object's velocity, where
the first value specifies the velocity in the x-direction and
the second in the y-direction [Default: c(0.0, 0.0)].}

\item{\code{active}}{A length-one logical vector indicating whether the object is
active [Default: TRUE].}
}}

\examples{
a <- new("Agent")
print(a)
a <- initialize(a,
                location = c(1.23, 10.5),
                velocity = c(5 * sqrt(2), 5 * sqrt(2)),
                active = FALSE)
validAgentObject(a)
print(a)
is.object(a)
isS4(a)

}
\seealso{
\code{\link{validAgentObject}}, \code{\link{initialize,Agent-method}}
}
