% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{getLocation}
\alias{getLocation}
\title{getLocation Generic}
\usage{
getLocation(object)
}
\arguments{
\item{object}{An object of class \linkS4class{Agent}}
}
\description{
Generic function definition for retreiving the location slot from an object
of class \linkS4class{Agent}.
}
