% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{methods}
\name{initialize,Agent-method}
\alias{initialize,Agent-method}
\title{Initialize an object of class "Agent".}
\usage{
\S4method{initialize}{Agent}(.Object, location = numeric(0),
  velocity = numeric(0), active = logical(0))
}
\arguments{
\item{.Object}{Name of an object of class \linkS4class{Agent}.}

\item{location}{A numerical vector of length 2 specifying the x- and
y-position of the object of class \linkS4class{Agent}. If not
specified the default location will be c(0, 0).}

\item{velocity}{A numerical vector of length 2 specifying the velocity in the
x- and y-direction of the object of class
\linkS4class{Agent}. If not specified the default velocity
will be c(0, 0).}

\item{active}{A length-one logical vector indicating whether the object of
class \linkS4class{Agent} is active (value: TRUE) or not
(value: FALSE). If not specified the default value will be TRUE.}
}
\description{
The initialize function sets the values for the slots of an object of class
\linkS4class{Agent}.
}
\examples{
a <- new("Agent")
print(a)
a <- initialize(a,
                location = c(1.23, 10.5),
                velocity = c(5 * sqrt(2), 5 * sqrt(2)),
                active = TRUE)
validAgentObject(a)
print(a)

}
\seealso{
\linkS4class{Agent}, \code{\link{Agent-class}},
\code{\link{validAgentObject}}
}
