% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{Lynx-class}
\alias{Lynx-class}
\alias{Lynx}
\title{An S4 class to represent a larger predator.}
\description{
An S4 class to represent a larger predator.
}
\section{Slots}{

\describe{
\item{\code{location}}{A numerical vector of length 2 specifying the x- and
y-position of the object [Default: c(0, 0)].}

\item{\code{velocity}}{A numeric vector for specifying the object's velocity, where
the first value specifies the velocity in the x-direction and
the second in the y-direction [Default: c(0.0, 0.0)].}

\item{\code{active}}{A length-one logical vector indicating whether the object is
active [Default: TRUE].}
}}

