% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\docType{methods}
\name{move}
\alias{move}
\alias{move,Agent-method}
\alias{move,Prey-method}
\alias{move,Bobcat-method}
\alias{move,Lynx-method}
\title{move Generic}
\usage{
move(object)

\S4method{move}{Agent}(object)

\S4method{move}{Prey}(object)

\S4method{move}{Bobcat}(object)

\S4method{move}{Lynx}(object)
}
\arguments{
\item{object}{An object of a defined S4 class (\linkS4class{Agent},
\linkS4class{Prey}, \linkS4class{Bobcat}. or \linkS4class{Lynx}).}
}
\description{
Generic function to move an object of a defined class.
}
\section{Methods (by class)}{
\itemize{
\item \code{Agent}: Move an object of class \linkS4class{Agent}.

\item \code{Prey}: Move an object of class \linkS4class{Prey}.

\item \code{Bobcat}: Move an object of class \linkS4class{Bobcat}.

\item \code{Lynx}: Move an object of class \linkS4class{Bobcat}.
}}

