% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\docType{methods}
\name{resetActivity}
\alias{resetActivity}
\alias{resetActivity,Agent,logical-method}
\alias{resetActivity,Agent,numeric-method}
\title{resetActivity Generic}
\usage{
resetActivity(object, value)

\S4method{resetActivity}{Agent,logical}(object, value)

\S4method{resetActivity}{Agent,numeric}(object, value)
}
\arguments{
\item{object}{An object of class \linkS4class{Agent}}

\item{value}{Either a length-one logical vector to reset the active slot of
the \code{\link{Agent-class}} object or a numerical vector of
length 2 to reset the velocity in the x- and y-direction of the
object of class \linkS4class{Agent}.}
}
\description{
Generic function to reset the activity of an object of class
\linkS4class{Agent}.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Agent,value = logical}: Reset the active slot to the specified value and reset the velocity slot to c(0, 0).

\item \code{object = Agent,value = numeric}: Reset the velocity slot to the specified values and reset the active slot to TRUE.
}}

