% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{setActive}
\alias{setActive}
\title{setAcitve Generic}
\usage{
setActive(object, active)
}
\arguments{
\item{object}{An object of class \linkS4class{Agent}}

\item{active}{A length-one logical vector indicating whether the object of
class \linkS4class{Agent} is active (TRUE) or not (FALSE)
[Default: TRUE].}
}
\description{
Generic function to set the active slot of an object of class
\linkS4class{Agent}.
}
