% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{setLocation}
\alias{setLocation}
\title{setLocation Generic}
\usage{
setLocation(object, location)
}
\arguments{
\item{object}{An object of class \linkS4class{Agent}}

\item{location}{A numerical vector of length 2 specifying the x- and
y-position of the object of class \linkS4class{Agent}
[Default: c(0, 0)].}
}
\description{
setLocation Generic
}
