% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{setVelocity}
\alias{setVelocity}
\title{setVelocity Generic}
\usage{
setVelocity(object, velocity)
}
\arguments{
\item{object}{An object of class \linkS4class{Agent}}

\item{velocity}{A numerical vector of length 2 specifying the velocity in the
x- and y-direction of the object of class \linkS4class{Agent}
[Default: c(0, 0)].}
}
\description{
setVelocity Generic
}
