% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\docType{methods}
\name{[,Agent,character-method}
\alias{[,Agent,character-method}
\title{Retrieve-method for slots}
\usage{
\S4method{[}{Agent,character}(x, i)
}
\arguments{
\item{x}{object of class "Agent"}

\item{i}{character, describing the slotname to retreive}
}
\description{
Extract parts of an object of class Agent by slot name.
}
