% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\docType{methods}
\name{[<-,Agent,character-method}
\alias{[<-,Agent,character-method}
\title{Replace-method for slots}
\usage{
\S4method{[}{Agent,character}(x, i) <- value
}
\arguments{
\item{x}{object of class "Agent"}

\item{i}{character, describing the slotname to be replace with a value}

\item{value}{ANY, value for the slotname specified by i to be replaced}
}
\description{
Replace values of slots in an object of class Agent by name
}
