% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validAgentObject.R
\name{validAgentObject}
\alias{validAgentObject}
\title{Check the validity of an object of class Agent.}
\usage{
validAgentObject(object)
}
\arguments{
\item{object}{An object of class \code{\link{Agent}}.}
}
\value{
TRUE or an error message.
}
\description{
The \code{validAgenObject} function checks whether an object of class
\code{\link{Agent}} is a valid object. An object of the \code{\link{Agent-class}}
is valid when:
\deqn{(object@velocity)^2 \le 100.0}
(in this case TRUE is returned), otherwise the message "The velocity is out
of bounds." will appear.
}
\examples{
a <- new("Agent")
print(a)
a <- initialize(a,
                location = c(1.23, 10.5),
                velocity = c(5 * sqrt(2), 5 * sqrt(2)),
                active = FALSE)
validAgentObject(a)
print(a)

}
\seealso{
\linkS4class{Agent}, \code{\link{Agent-class}},
\code{\link{initialize,Agent-method}}
}
