% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validityFunctions.R
\name{validBobcatObject}
\alias{validBobcatObject}
\title{Check the validity of an object of class Bobcat.}
\usage{
validBobcatObject(object)
}
\arguments{
\item{object}{An object of class \code{\link{Bobcat}}.}
}
\value{
TRUE or an error message.
}
\description{
The \code{validBobcatObject} function checks whether an object of class
\code{\link{Bobcat}} is a valid object. An object of the
\code{\link{Bobcat-class}} is valid when:
\deqn{(object@velocity)^2 \le 85.0}
(in this case TRUE is returned), otherwise the message "The velocity is out
of bounds." will appear.
}
\seealso{
\linkS4class{Bobcat}, \code{\link{Bobcat-class}},
}
