% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validityFunctions.R
\name{validLynxObject}
\alias{validLynxObject}
\title{Check the validity of an object of class Bobcat.}
\usage{
validLynxObject(object)
}
\arguments{
\item{object}{An object of class \code{\link{Lynx}}.}
}
\value{
TRUE or an error message.
}
\description{
The \code{validLynxObject} function checks whether an object of class
\code{\link{Lynx}} is a valid object. An object of the \code{\link{Lynx-class}}
is valid when:
\deqn{(object@velocity)^2 \le 95.0}
(in this case TRUE is returned), otherwise the message "The velocity is out
of bounds." will appear.
}
\seealso{
\linkS4class{Lynx}, \code{\link{Lynx-class}},
}
