% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validityFunctions.R
\name{validPreyObject}
\alias{validPreyObject}
\title{Check the validity of an object of class Prey.}
\usage{
validPreyObject(object)
}
\arguments{
\item{object}{An object of class \code{\link{Prey}}.}
}
\value{
TRUE or an error message.
}
\description{
The \code{validPreyObject} function checks whether an object of class
\code{\link{Prey}} is a valid object. An object of the \code{\link{Prey-class}}
is valid when:
\deqn{(object@velocity)^2 \le 70.0}
(in this case TRUE is returned), otherwise the message "The velocity is out
of bounds." will appear.
}
\seealso{
\linkS4class{Prey}, \code{\link{Prey-class}},
}
