% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{Agent-class}
\alias{Agent-class}
\alias{Agent}
\title{An S4 class to represent an agent.}
\description{
An S4 class to represent an agent.
}
\section{Slots}{

\describe{
\item{\code{location}}{A numeric vector for specifying the object's location, where
the first value specifies x and the second y  [Default:
c(0.0, 0.0)].}

\item{\code{velocity}}{A numeric vector for specifying the object's velocity, where
the first value specifies the velocity in the x-direction and
the second in the y-direction [Default: c(0.0, 0.0)].}

\item{\code{active}}{A logical vector indicating whether th object is active
[default: TRUE].}
}}

\examples{
a <- new("Agent")
print(a)
a <- initialize(a,
                location = c(1.23, 10.5),
                velocity = c(5 * sqrt(2), 5 * sqrt(2)),
                active = TRUE)
validAgentObject(a)
print(a)

}
\seealso{
\code{\link{validAgentObject}}, \code{\link{initialize,Agent-method}}
}
